feat: Pagination

+++ reviews/views.py
@@ -3,27 +3,35 @@ from django.shortcuts import get_object_or_404, render
 from rest_framework.views import APIView, Response, status
 from rest_framework.authentication import TokenAuthentication
 from rest_framework.permissions import IsAuthenticatedOrReadOnly
+from rest_framework.pagination import PageNumberPagination
 
-from movies.models import Movie
+from users.permissions import IsAdminOrCritic
 
 from .models import Review
 from .serializers import ReviewSerializer
 
 
-class ReviewView(APIView):
+class ReviewView(APIView, PageNumberPagination):
     def get(self, request):
         reviews = Review.objects.all()
-        serializer = ReviewSerializer(reviews, many=True)
-        return Response(serializer.data)
+        result_page = self.paginate_queryset(reviews, request, self)
+        serializer = ReviewSerializer(result_page, many=True)
+        return self.get_paginated_response(serializer.data)
+
 
 class ReviewViewDetail(APIView):
-    #Somente admin ou próprio crítico que fez a review
+    authentication_classes = [TokenAuthentication]
+    permission_classes = [IsAdminOrCritic]
+    # Somente admin ou próprio crítico que fez a review
+
     def delete(self, request, review_id):
         review = get_object_or_404(Review, pk=review_id)
+        self.check_object_permissions(request, review)
         review.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
-class MovieReviewViewDetail(APIView):
+
+class MovieReviewViewDetail(APIView, PageNumberPagination):
     authentication_classes = [TokenAuthentication]
     permission_classes = [IsAuthenticatedOrReadOnly]
 
@@ -35,12 +43,9 @@ class MovieReviewViewDetail(APIView):
 
     def get(self, request, movie_id):
         try:
-            reviews = Review.objects.get(pk=movie_id)
-            serializer = ReviewSerializer(reviews)
-            return Response(serializer.data)
+            reviews = Review.objects.filter(movie_id=movie_id)
+            result_page = self.paginate_queryset(reviews, request, self)
+            serializer = ReviewSerializer(result_page, many=True)
+            return self.get_paginated_response(serializer.data)
         except Review.DoesNotExist:
             return Response({"message": "Review not found."}, status.HTTP_404_NOT_FOUND)
-
-    
-
-
